新增单元测试

+++ FFDBTests/TestModel.h
@@ -12,5 +12,5 @@
 
 @property(nonatomic,copy) NSString *name;
 @property(nonatomic,copy) NSString *memory;
-@property(nonatomic,assign) float time;
+@property(nonatomic,assign) double time;
 @end
